Increment the API version, switch EnergyValue to only accept a float (over Number, small performance increase), and better protection of learning Ores

+++ src/main/java/com/pahimar/ee3/command/CommandSetEnergyValue.java
@@ -51,12 +51,12 @@ public class CommandSetEnergyValue extends CommandBase
         else
         {
             Item item = getItemByText(commandSender, args[2]);
-            double energyValue = 0;
+            float energyValue = 0;
             int metaData = 0;
 
             if (args.length >= 4)
             {
-                energyValue = parseDoubleWithMin(commandSender, args[3], 0);
+                energyValue = (float) parseDoubleWithMin(commandSender, args[3], 0);
             }
             else if (args.length >= 5)
             {
